pyproject.toml
[tool.poetry]
name = "resting"
version = "0.1.0"
description = "Simple REST test tool"
authors = ["Konstantin Vinogradov <kostya.vinogradov@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Pygments = "^2.11.2"
aiohttp = "^3.8.1"
PyYAML = "^6.0"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
black = "^22.1.0"
types-Pygments = "^2.9.16"
types-PyYAML = "^6.0.4"

[tool.poetry.scripts]
resting-script = "resting.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
